fix(be): fix noimage cocktail delete error

+++ backend/src/routes/recipe.js
@@ -8,6 +8,7 @@ import axios from 'axios';
 import multer from 'multer';
 import fs from 'fs';
 import { Sequelize } from 'sequelize';
+import { logger } from '../../winston/winston';
 
 dotenv.config();
 const router = express.Router();
@@ -113,12 +114,13 @@ router.get('/:cocktailId', checkAccess, async (req, res) => {
       color: [],
       ingred: [],
       instruction: '',
+      image: null
     };
 
     // cocktail get
     const cocktailId = req.params.cocktailId;
     const cocktail = await db.COCKTAIL.findByPk(cocktailId, {
-      attributes: ['CNO', 'NAME', 'ALCOHOLIC', 'INSTRUCTION'],
+      attributes: ['CNO', 'NAME', 'ALCOHOLIC', 'INSTRUCTION', 'IMAGE_PATH'],
       include: [
         {
           model: db.COLOR,
@@ -142,6 +144,7 @@ router.get('/:cocktailId', checkAccess, async (req, res) => {
     data.name = cocktail.NAME;
     data.alcoholic = cocktail.ALCOHOLIC;
     data.instruction = cocktail.INSTRUCTION;
+    data.image = cocktail.INAGE_PATH
 
     // color
     for (let i = 0; i < color.length; i++) {
@@ -268,16 +271,17 @@ router.delete('/:cocktailId', async (req, res) => {
     const cocktail = await db.COCKTAIL.findByPk(cocktailId);
 
     // file system에서 이미지파일 삭제
-    const oldFilePath = `./${cocktail.IMAGE_PATH}`;
-    console.log(oldFilePath);
-    fs.unlinkSync(oldFilePath);
-    console.log(cocktail);
+    if (cocktail.IMAGE_PATH) {
+      fs.unlinkSync(oldFilePath);
+      const oldFilePath = `./${cocktail.IMAGE_PATH}`;
+      console.log(oldFilePath);
+    }
     await cocktail.destroy();
     // color, recipe - CASCADE TRIGGER로 자동 삭제
-
+    logger.info('Cocktail delete 성공');
     res.status(200).json({ success: true, message: 'Cocktail delete 성공' });
   } catch (error) {
-    console.log(error);
+    logger.error(error);
     return res
       .status(400)
       .json({ success: false, message: 'Cocktail delete 실패', error });
@@ -501,9 +505,9 @@ router.get('/detail/review/:cocktailId', checkAccess, async (req, res) => {
           model: db.USER,
           as: 'UNO_USER',
           required: true,
-          attributes: ['NICKNAME', 'LEVEL']
-        }
-      ]
+          attributes: ['NICKNAME', 'LEVEL'],
+        },
+      ],
     });
     console.log(post.count);
     console.log(post.rows);
@@ -527,7 +531,7 @@ router.get('/image/:cocktailId', async (req, res) => {
     const cocktail = await db.COCKTAIL.findByPk(cocktailId);
     // 이미지 없으면 로고 이미지 보내줌
     if (!cocktail.IMAGE_PATH) {
-      const data = fs.readFileSync("uploads/cocktailImage/logo.png");
+      const data = fs.readFileSync('uploads/cocktailImage/logo.png');
       res.writeHead(200, { 'Content-Type': 'image/jpg' }); //보낼 헤더를 만듬
       res.write(data);
       return res.end();
